Migrate from Turso/libsql to Cloudflare D1 and fix cssom crash
Switch database from Turso (libsql over HTTP) to Cloudflare D1 binding,
remove libsql/cross-fetch shims, upgrade vinext to 0.0.18 (dropping patch),
add optimizeDeps for cssom to fix Vite CJS module runner crash, and remove
unused @mdx-js/mdx dependency.
