+++ drizzle-kit.config.ts
@@ -1,0 +1,10 @@
+import type { Config } from "drizzle-kit";
+
+export default {
+	schema: "./schema.ts",
+	out: "./migrations",
+	dialect: "sqlite",
+	dbCredentials: {
+		url: "file:database.db",
+	},
+} satisfies Config;
