+++ env.ts
@@ -1,8 +1,6 @@
 import { z } from "zod";
 
 const envSchema = z.object({
-	DATABASE_URL: z.string().min(1),
-	DATABASE_AUTH_TOKEN: z.string().default(""),
 	GITHUB_CLIENT_ID: z.string().min(1),
 	GITHUB_CLIENT_SECRET: z.string().min(1),
 	NODE_ENV: z.enum(["development", "production", "test"]).default("development"),
