+++ package.json
@@ -7,9 +7,7 @@
 		"blog": "bun run cli/blog.ts",
 		"blog:prod": "BLOG_API_URL=https://solberg.is bun run cli/blog.ts",
 		"atlas": "./scripts/atlas-diff.sh",
-		"dev": "concurrently \"bun run dev:next\" \"bun run dev:db\" \"bun run studio\"",
-		"dev:next": "vinext dev",
-		"dev:db": "turso dev --db-file database.db --port 9797",
+		"dev": "vinext dev",
 		"build": "vite build",
 		"start": "vinext start",
 		"deploy": "vite build && wrangler deploy",
@@ -21,17 +19,15 @@
 		"format-and-lint:fix": "bun run lint:fix && bun run format",
 		"knip": "dotenvx run -- knip",
 		"knip:fix": "dotenvx run -- knip --fix",
-		"generate:migration": "drizzle-kit generate",
-		"run:migration": "drizzle-kit up",
-		"studio": "drizzle-kit studio",
+		"generate:migration": "drizzle-kit generate --config=drizzle-kit.config.ts",
+		"run:migration": "wrangler d1 execute solberg-blog --file",
+		"studio": "drizzle-kit studio --config=drizzle-kit.config.ts",
 		"postinstall": "lefthook install"
 	},
 	"dependencies": {
 		"@headlessui/react": "^2.2.9",
 		"@heroicons/react": "^2.2.0",
 		"@hono/zod-validator": "^0.7.6",
-		"@libsql/client": "^0.17.0",
-		"@mdx-js/mdx": "^3.1.1",
 		"@monaco-editor/react": "^4.7.0",
 		"@tanstack/react-pacer": "^0.19.1",
 		"arctic": "^3.7.0",
@@ -55,10 +51,9 @@
 		"tailwind-merge": "^3.4.0",
 		"tailwind-variants": "^3.2.2",
 		"tailwindcss-react-aria-components": "^2.0.1",
-		"turso": "^0.1.0",
 		"tw-animate-css": "^1.4.0",
 		"usehooks-ts": "^3.1.1",
-		"vinext": "^0.0.15",
+		"vinext": "^0.0.18",
 		"zod": "^4.3.5"
 	},
 	"devDependencies": {
@@ -71,7 +66,6 @@
 		"@types/react-dom": "^19.2.3",
 		"@types/rss": "^0.0.32",
 		"@vitejs/plugin-rsc": "^0.5.21",
-		"concurrently": "^9.1.0",
 		"drizzle-kit": "^0.31.8",
 		"knip": "^5.80.2",
 		"lefthook": "^2.0.14",
@@ -92,7 +86,5 @@
 			"esbuild"
 		]
 	},
-	"patchedDependencies": {
-		"vinext@0.0.15": "patches/vinext@0.0.15.patch"
-	}
+	"patchedDependencies": {}
 }
